feat(pic8): support images up to 2^31 in size
Don't silently truncate images to 2^16.

+++ pic8.cpp
@@ -13,8 +13,8 @@ void pic8::allocate(int w, int h) {
         internal_error("pic8 already allocated!");
         return;
     }
-    width = short(w);
-    height = short(h);
+    width = w;
+    height = h;
     if (w % 4) {
         w += 4 - w % 4;
     }
